backend/requirements.txt
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
distro==1.9.0
docker==7.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
Flask==3.0.2
Flask-Cors==4.0.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==3.2.0
mangum==0.17.0
MarkupSafe==2.1.5
packaging==24.0
paramiko==3.4.0
postgres==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pycparser==2.22
PyNaCl==1.5.0
pyrsistent==0.20.0
python-dotenv==1.0.1
pywin32==306
PyYAML==5.4.1
requests==2.31.0
six==1.16.0
texttable==1.7.0
typing_extensions==4.10.0
urllib3==2.2.1
websocket-client==0.59.0
Werkzeug==3.0.2
zipp==3.18.1
